pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imglabel"
version = "0.1.0"
description = "A tool for labeling paint damage patches in images."
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "matplotlib",
    "numpy",
    "Pillow",
    "scipy",
    "scikit-image",
    "PyYAML",
    "PyQt6",
]

[project.scripts]
imglabel = "imglabel.cli:main"
imglabel-gui = "imglabel.gui:main"

[tool.hatch.build.targets.wheel]
packages = ["src/imglabel"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "C90"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
